Add UI of Status and add script of 'KillEnemyCount'
Added a UI feature where every time a player kills an enemy of the current type, the progress bar advances by one-third. When the player kills three of the same enemies, the text 'Press 3' pops up. Additionally, integrated a 'KillEnemyCount' to track the count and display it on the UI.